+++ shared/number.py
@@ -1,4 +1,4 @@
-def is_int(num):
+def can_be_int(num):
     try:
         int(num)
 
@@ -7,11 +7,11 @@ def is_int(num):
         return False
 
 
-def is_not_int(num):
-    return not is_int(num)
+def cant_be_int(num):
+    return not can_be_int(num)
 
 
-def is_float(num):
+def can_be_float(num):
     try:
         float(num)
 
@@ -20,14 +20,14 @@ def is_float(num):
         return False
 
 
-def is_not_float(num):
-    return not is_float(num)
+def cant_be_float(num):
+    return not can_be_float(num)
 
 
-def is_int_or_float(num):
-    return is_int(num) and is_float(num)
+def can_be_int_or_float(num):
+    return can_be_int(num) or can_be_float(num)
 
 
-def is_not_int_or_float(num):
-    return not is_int_or_float(num)
+def cant_be_int_or_float(num):
+    return not can_be_int_or_float(num)